refactor: geometry tools
